Rewrite IonizationTunnel and derivatives to use templates
Rewrite IonizationTunnel and the new models introduced by Mironov to be
more compact and use templates.
